move valitadion logic to validator class, move period list publishing to own class
