+++ ClubManagerWeb/src/main/java/com/infoshareacademy/niewiem/reservations/validators/ReservationValidator.java
@@ -2,6 +2,7 @@ package com.infoshareacademy.niewiem.reservations.validators;
 
 import com.infoshareacademy.niewiem.reservations.enums.Period;
 import com.infoshareacademy.niewiem.shared.validators.GenericValidator;
+import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -12,7 +13,16 @@ import java.util.List;
 public class ReservationValidator extends GenericValidator {
     private static final Logger LOG = LoggerFactory.getLogger(ReservationValidator.class);
 
-    public boolean validatePeriodExists(String periodParam, List<String> errors) {
+    public boolean validatePeriodParam(String periodParam, List<String> errors) {
+        if (StringUtils.isEmpty(periodParam)) {
+            LOG.info("No period parameter in request.");
+            return false;
+        }
+
+        return validatePeriodExists(periodParam, errors);
+    }
+
+    private boolean validatePeriodExists(String periodParam, List<String> errors) {
         for (Period p : Period.values()) {
             if (p.name().equalsIgnoreCase(periodParam)) {
                 LOG.debug("Period option exists. ({})", periodParam);
